yoti-sdk-api/src/main/java/com/yoti/json/ResourceMapper.java
package com.yoti.json;

import java.nio.charset.StandardCharsets;
import java.text.SimpleDateFormat;
import java.util.Locale;
import java.util.TimeZone;

import com.fasterxml.jackson.annotation.JsonAutoDetect;
import com.fasterxml.jackson.annotation.JsonSetter;
import com.fasterxml.jackson.annotation.Nulls;
import com.fasterxml.jackson.core.JsonProcessingException;
import com.fasterxml.jackson.databind.DeserializationFeature;
import com.fasterxml.jackson.databind.ObjectMapper;
import com.fasterxml.jackson.databind.cfg.MapperConfig;
import com.fasterxml.jackson.databind.introspect.VisibilityChecker;
import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;

public final class ResourceMapper {

    private static final ObjectMapper MAPPER = new ObjectMapper();

    static {
        MAPPER.disable(DeserializationFeature.WRAP_EXCEPTIONS)
                .disable(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES)
                .disable(DeserializationFeature.READ_DATE_TIMESTAMPS_AS_NANOSECONDS)
                .setVisibility(configureVisibility(MAPPER.getDeserializationConfig()))
                .setDefaultSetterInfo(JsonSetter.Value.forValueNulls(Nulls.SKIP))
                .setDateFormat(utc())
                .registerModule(new JavaTimeModule());
    }

    private static SimpleDateFormat utc() {
        SimpleDateFormat format = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'", Locale.ROOT);
        format.setTimeZone(TimeZone.getTimeZone("UTC"));

        return format;
    }

    private ResourceMapper() { }

    private static VisibilityChecker<?> configureVisibility(MapperConfig<?> config) {
        return config.getDefaultVisibilityChecker()
                .withFieldVisibility(JsonAutoDetect.Visibility.NONE)
                .withGetterVisibility(JsonAutoDetect.Visibility.NONE)
                .withIsGetterVisibility(JsonAutoDetect.Visibility.NONE)
                .withSetterVisibility(JsonAutoDetect.Visibility.NONE)
                .withCreatorVisibility(JsonAutoDetect.Visibility.NONE);
    }

    public static byte[] writeValueAsString(Object object) throws JsonProcessingException {
        return MAPPER.writeValueAsString(object).getBytes(StandardCharsets.UTF_8);
    }

}
